requirements.txt
streamlit
opencv-python-headless
Pillow
numpy
ultralytics
cvzone
sort
matplotlib
fpdf
